+++ Sources/Requests/Requests+SetupOrgReq.swift
@@ -1,22 +1,31 @@
 import Foundation
 
 extension Requests {
-    public struct SetOrgTypeReq: Codable, Hashable, Sendable {
+    public struct SetupOrgReq: Codable, Hashable, Sendable {
         public let orgType: OrgTypeEnum
+        /// Must be a string starting with `@`. May only contain a-z, A-Z, 0-9, _, -. May not be longer than 30 characters.
+        public let handle: String?
+        public let companyName: String?
         /// Additional properties that are not explicitly defined in the schema
         public let additionalProperties: [String: JSONValue]
 
         public init(
             orgType: OrgTypeEnum,
+            handle: String? = nil,
+            companyName: String? = nil,
             additionalProperties: [String: JSONValue] = .init()
         ) {
             self.orgType = orgType
+            self.handle = handle
+            self.companyName = companyName
             self.additionalProperties = additionalProperties
         }
 
         public init(from decoder: Decoder) throws {
             let container = try decoder.container(keyedBy: CodingKeys.self)
             self.orgType = try container.decode(OrgTypeEnum.self, forKey: .orgType)
+            self.handle = try container.decodeIfPresent(String.self, forKey: .handle)
+            self.companyName = try container.decodeIfPresent(String.self, forKey: .companyName)
             self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
         }
 
@@ -24,11 +33,15 @@ extension Requests {
             var container = encoder.container(keyedBy: CodingKeys.self)
             try encoder.encodeAdditionalProperties(self.additionalProperties)
             try container.encode(self.orgType, forKey: .orgType)
+            try container.encodeIfPresent(self.handle, forKey: .handle)
+            try container.encodeIfPresent(self.companyName, forKey: .companyName)
         }
 
         /// Keys for encoding/decoding struct properties.
         enum CodingKeys: String, CodingKey, CaseIterable {
             case orgType = "org_type"
+            case handle
+            case companyName = "company_name"
         }
     }
 }